// ClientApp/src/store/aicompletion/aicompletion.types.ts
export enum AICOMPLETION_ACTION_TYPES {
    CREATE_START = 'aicompletion/CREATE_START',
    CREATE_SUCCESS = 'aicompletion/CREATE_SUCCESS',
    CREATE_FAILED = 'aicompletion/CREATE_FAILED',
    UPDATE_START = 'aicompletion/UPDATE_START',
    UPDATE_SUCCESS = 'aicompletion/UPDATE_SUCCESS',
    UPDATE_FAILED = 'aicompletion/UPDATE_FAILED',
    DELETE_START = 'aicompletion/DELETE_START',
    DELETE_SUCCESS = 'aicompletion/DELETE_SUCCESS',
    DELETE_FAILED = 'aicompletion/DELETE_FAILED',
    FETCH_ALL_START = 'aicompletion/FETCH_ALL_START',
    FETCH_ALL_SUCCESS = 'aicompletion/FETCH_ALL_SUCCESS',
    FETCH_ALL_FAILED = 'aicompletion/FETCH_ALL_FAILED'
};

export type AiCompletion = {
    
}